Add abc encode/decode functionality  (bug 400055, r=jodyer)
Adds basic functionality to encode a .abc file into a human readable .asm file
and vice versa.

There is a verbose and non-verbose mode for encoding.  If verbose is selected
then most of the opcode operands are rendered in a longer more human readable
format (e.g. strings are printed as strings, for names the entire qualified
name is displayed).  In non-verbose mode all the operands are simply printed as
indexes to the appropriate constant pool.

The decoder can construct an abc from either format.  This makes it easier to
edit the bytecode in verbose mode, as you could just change a string inline,
instead of having to add an entry to the constant pool, calculate its index,
etc.

Also, modified the ABCBytestream class to provide read and write functionality
for common abc constructs (like variable length inetgers).