Add support for detecting imminent buffer overruns for our generated (JIT'd) code.
Currently if we do not allocate a large enough buffer for
code generation we run the risk of overrunning the buffer
since there are no downstream checks.  This issue was
exacerbated by a change made to getMDBuffer() which attempts
to get a smaller and smaller buffer when an allocation fails.
Thus, increasing the chance of getting a buffer which is too
small.  Previously we attempted to allocate a large enough
buffer (size estimate based on largest possibly needed) and
failed gracefully if the memory could not be obtained, thus
it was somewhat safer to carry on assuming that the buffer
was of sufficient capacity.

This change checks that during code generation we do not
run off the end of the buffer.   It is similar to the check
that exists when the buffer guard feature is disabled except
that we are not explicitly growing the buffer.