got a bunch of stuff working
added LIR_calli, fcalli

indirect calls have one additional arg that is the expression of
the address (usually a load).  CallInfo.count_args() and get_sizes()
take this into account

when calling an IMT thunk there's a forth argument that's the iid
(interface id).  so we have CallInfo signatures for both.  we will
need special smarts in the assembler to deal with this funny calling
convention.  (maybe want to shuffle the argument to a different spot
and possibly encode IMT calls using a calling convention rather
than just having an extra random arg).