Allow NJ to expand the cache as needed
During integration of NJ to TC we copied the allocation
pattern used in TC, namely we created one large buffer
per pool.   Unfortunately in NJ's case the memory is not
a virtual alloc but is fully committed.  Also the current
page tracking code in Fragmento hits each page as
its allocated.  A disproportionate amount of time was
thus being spent in this area.  Short running benchmarks
were especially sensitive.

Quick fix is to allow NJ to alloc mem as needed.