improve register allocation at forward branches (union vs intersect)
mergeRegisterState()'s merge policy is as follows:

current == saved     skip
current & saved      evict current
current & !saved     evict current
!current & saved     assign saved

this means if the saved state is empty, the current state will
be cleared out by spilling whatever is live.  This is what you
want when merging a side exit or the code after a loop, so you
dont pollute the current state with "cold" register assignments.

In the common case of a forward branch, however, we want to keep
registers live in the current state, if possible.  I renamed
mergeRegisterState() to intersectRegisterState(), and added a new
unionRegisterState() method whose policy is:

current == saved     skip
current & saved      evict current
current & !saved     KEEP CURRENT   <--- changed
!current & saved     assign saved

In the very common case of a null pointer check, the target state
is empty, and this policy means we dont spill everything live just
because of a null pointer check.

This reduced the overall instruction count in boidshack from 6624 to 5907.