Keep x87 stack balanced, eliminate redundant loads in CopyPropation, more const folding
at the point of a forward branch we need to evict x87 st(0) if it is not
loaded at the target of the branch.

copy propagation was only using stores to initialize state, but loads can
too, especially to handle eliminating redundant loads with no intermediate stores.

added const folding for integer unary ! - and binary & | ^ + - *