improved prolog, add stk overflow check, add LoadFilter
saved-register spilling is now done by using one LIR_param for each
saved register, and ensuring that procedure exit paths (LIR_ret)
assign those params to the appropriate regs.  They will be the lowest
priority registers, first to be spilled if they are needed, but for
simple functions, they may not be needed and won't get spilled.

LoadFilter eliminates a second load if it's a copy of an earlier load
and no intervening stores have occurred.

Stack overflow checks precede each prolog, like this:

   cmp esp, XXX  // bottom limit of stack
   jnb skip
   mov ecx, 4(esp) // env
   call stkover
   skip:
   ...

this is simpler than MIR, which goes out of its way to ensure the
forward branch is staticly predicted not taken by putting the call
to stkover after the end of the method.  prediction makes no difference
in testing, so this is simpler.

this patch removes the dynamic alignment of ESP, if we need it
for TT or TM we can easily add it back in.  this just makes it
similar to MIR.   when the alignment is not needed, doing it anyway
is a measureable slowdown.