Support for generating pictures of control-flow graphs of JIT'd code.
Using the -bbgraph option on the command-line results in a Graphviz
dot compatible description to be generated and output to the terminal.

This output can then be cut & pasted into a text file and a command
like 'dot -Tjpg < infile' will render the graph into a JPEG.

The component that builds the graph, BlockLocator, is implemented
as a LirWriter, meaning it can be placed in stream and is thus dependent
only on the input LIR stream.  Also branches need not be fully resolved
while the stream is processing.  A final step, executed when fin() is
called, resolves all these previously un-targetted branches.