Fixed bug causing too much spilling, other arm tweaks
Assembler.findRegFor2() contained a fix for the case when the rhs
parameter was already assigned a register, but the register wasn't
in the allow mask.

That fix interacted with the code to prevent the lhs parameter from
stealing the rhs parameter's register.  The interaction caused
the RHS register to spill unecessarily when it already *was* allowed.

The new fix avoids both issues.

Other:
* add arm assm to GC::cleanStack()
* fix gcc error in GCHeapUnix.cpp
* move Config.runmode out of JIT ifdefs, to enable non-jit configs to build
* factor LIR_ret impl code out of Assembler, into Nativei386.cpp
* removed NJ_STACK_OFFSET which was =0 on all supported platforms
* move asm_pusharg into cpu-specific files
* memset RegAlloc.usepri[] at init time
* RegAlloc.active[] and usepri[] must be the same size