implement describeType() (and a few related calls) for TC/TR.
This is basically a port of the version implemented for TT, with a few bug fixes and tweaks -- the native API produces the type info as a JSON object, and AS3 code transforms it into XML.

avmplus.describeType() is essentially a superset of flash.utils.describeType() -- it produces the same results, but with some options to allow skipping the generation of info that isn't necessary for a given call. (Exception: it doesn't provide the "alias" attribute as that is a Flash-specific concept that is unlikely to be moved to the VM anytime soon)

Also found & fixed a few bugs related to TraitsMetadata caching (we hadn't been exercising every possible use, as it turns out)... some results were wrong, and a few could result in crashes.