Redo the native-thunk generation code yet again: some recalcitrant compilers are failing (or crashing!) on the AVMTHUNK_CALL_FUNCTION_x macros, but can apparently handle the same code in smaller pieces, so we now generate explicit code to separately extract the method from the MethodEnv, cast it appropriately, and call it. In so doing, the macro above went away entirely, as the new code (hopefully) uses techniques that will work on all compilers.
I also added a new option to nativegen.py, --uniquethunk, which generates a unique thunk for every native method and calls the implementation method directly, avoiding the casting game entirely. This isn't in use by default, as it can lead to serious code bloat for clients with lots of native methods, but I decided to leave it in as it might prove useful in the future.