Enable ARM jit (LIR), off by default (r=rreitmai+, vlad@mozilla+)
Big cleanup of opcode macros, better use of the ARM instruction set, removed
a lot of x86 holdovers from the code.

No VFP support yet, and not yet passing all the sanities.  It's enabled
here to get into the release stream for testing in other build environments
and as a checkpoint.

Use the -Ojit option to enable ARM compilation for all methods, otherwise
the interpreter is used for all methods.