Revise functions-names to use minimal runtime storage. (r=lhansen, edwsmith)
Instead of calculating the fancy name at parse time, we now store just an index to the base name (or qname in the case of methods), then construct the decorated name on demand. This reduces the per-AbstractFunction size, plus eliminates a lot of rarely-used strings. The only extra memory we use now is a per-pool list of ints (1 per function).

Feature is now controlled by VMCFG_METHOD_NAMES, which is enabled by default if DEBUGGER||AVMPLUS_VERBOSE... but also by a runtime flag, config.methodNames, so that clients that don't care about accurate methodnames can disable it to save further memory.