Remove abcasm dependency on asc.jar (bug 480304 r=lhansen+)
We just needed the constants.  Not depending on asc.jar means we don't
need to use the -a switch to abcasm.sh, and means that java -jar invocation
should work now (although abcasm.sh still uses classpath invocation).

Also squeezed in support for the sign-extend opcodes sxi1, sxi8, and sxi16,
in the grammar and in test/acceptance/abcasm/mops.abs