Factor verifyall code out of the Verifier body, into a CodeWriter filter (bug 481244, r=jodyer)
VerifyallFilter is pushed on to the coder pipeline if verifyall mode is enabled.

Use overloaded convenience methods instead of default args, to reduce
liklihood of a CodeWriter filter adding a virtual method with the wrong signature,
and to reduce code size when the helpers are not inlined.  (call to helper with
one less arg is smaller than call to writeXXX with null pointer implicitly passed).