Major changes to MethodInfo to reduce memory usage. (r=edwsmith)
Primarily, a number of fields that can be reconstituted from re-parsing ABC are
now split and cached in a new MethodSignature struct. This exactly analogous to
the Traits-TraitsBinding relationship. Testing on decent-sized Flex apps shows
a high-water memory use reduction of 300k-500k when used with a cache size of
32.

A bit more testing is necessary prior to pushing, primarily in the performance
area, but initial testing has shown performance to be fine so I'm offering this
patch for initial review (pending further perf testing).

I also changed the default size of the TraitsBindings cache to 32 -- it was
formerly unlimited, but it seems clear at this point that 32 is a better
default.

Also rewrote methodInfo::boxArgs to use BuiltinType instead of repeated
if-else.