bug 482478: add differential testing script, avmdiff.py, runs against acceptance test suite Differential testing runs multiple shells and runtime flags against abc testcases comparing the output and exit codes. Any differences are flagged as failures.  The intent is to run acceptance tests substuting the shell with the avmdiff.py script.
The configuration script by default is avmdiff.cfg.  The configuration contains sections [builds],[options],[combined].  The builds
is a list of vm shells.  The options is a list of runtime flags.  Each build is run with each option.  The combined is a list of shell
and flag combinations.  The combined section is used when only a specific shell has a runtime flag available.

The differential testing script will be used with buildbot to automate running acceptance tests with several flags.