Toplevel no longer a ScriptObject (r=edwsmith)
Toplevel is now no longer a ScriptObject, but just a plain GCObject. (It holds
a reference to the toplevel ScriptObject.)

Also added Toplevel::findClassInPool so subclasses don't have to replicate the
logic.

Also privatized the VTable::toplevel field.

Also fix a minor bug in avmplusList<> (dtor needed to clear the gc field).