Remove the redundant MethodEnv::declTraits declaration to conserve memory. (r=edwsmith)
This was added to detect certain pathological ABC/SWF files, that call newfunction multiple times on the same method, but with different scopes; since the JIT emits code that assumes the scope in effect, it could lead to crashes. The original fix was to record the declaringTraits from the MethodInfo in the MethodEnv as well, and if it ever differs, fail: this had the effect of allowing the malformation as long as earlier incarnations of the function were never called again.

The revised fix relies on the fact that it's really the ScopeTypeChain (not the Traits) that we care about, and we already have access to those without adding an extra field to MethodEnv (which we allocate many thousands of, thus size matters).

The only gotcha here is that activationTraits never bothered to set its ScopeTypeChain properly prior to this (it didn't need to, it was able to rely on its ScopeChain being correct), but this was a trivial fix.

Also corrected an apparent oddity in the creation of the ScopeChain for Class instances (in MethodEnv::newclass)... we had been using the scope for Class$ (cscope) for Class (iscope) as well, but this did not agree with the ScopeTypeChain that Class was expecting. Using the conventionally-generated iscope makes this agree.

Also added a necessary clearing of the MethodSignature cache to makeIntoPrototypeFunction (fairly harmless but was generating an assertion).

Also added an acceptance test for this malformed case, using abcasm to generate the questionable ABC sequence.