bug 491724: add windows pyspy test to cmdline testsuite r+brbaker bug 491773: improve smoke test result output r+brbaker
pyspy test runs an abc and calls pyspy using windows named sockets to trigger
memory profile dump to stdout.  the cmdline test verifies the memory dump
triggers and some basic stats are shown in stdout. pyspy tests only run on
python.org version of python and requires the win32 extension python library
to be installed.

smoke tests are displayed with env variables not expanded.  if a test fails
the stdout is shown at the bottom of the results.