Adds free block coalescing to the allocator.
The CodeList (block header) structure now contains references to its adjacent
neighbors as well as a next pointer for building lists of blocks:

    |block|---higher-->|block|
    |     |<--lower----|     |

lower is null at the low-end and higher is null at the high end.  Each page
allocated from GCHeap has a small block at the end which is always in use and
is a teriminator of the linked list of blocks.

The heapblocks array is now a linked list using the terminal blocks of each
heap allocation.

CodeAlloc.freelist is no longer used; each block has a status (free or in-use)
and we locate free blocks by searching the heapblocks list.  No attempt has
been made to make allocation fast since it hasn't shown up as a hotspot.

Also, the block header (CodeList) is bigger that necessary, in particular the
isFree bit could be stuffed into one of the pointer fields.  Since we don't
allocate more than 4K at a time, we could also combine lower and higher/end
into two int16 size fields (allowing up to 64k per block on 32bit systems).
This would make CodeList be 2 words instead of 4.

also:
* commented out debugging memset
* CodeAlloc is constructed with GCHeap* now since it no longer has
  an internal array which requires GC* (previous review)