Grab bag of fixes from Box porting work: (r=lhansen, edwsmith):
-- getdescendantslate is never called. It needs to die.

-- many (native) setters are declared as returning "undefined" in as3 but return
"void" in C++... leading to unpredicable results in the result-code register in
obscure situations. this patch enforces that the expected C++ signature is
"void" regardless of the as3 signature.

-- CallStackNode::boxed is unused and should be removed. Also added the opaque
"FramePtr" type for CSN's framep field; it was formerly typed as Atom* but this
is misleading as the type is really Atom (if interpreted) or native-rep (if
jitted).

-- VectorClass inits various "empty" fields to NULL, which is not (technically) a
legal Atom value. We should init these areas to nullObjectAtom instead. (This
wasn't causing any problem I'm aware of, but can lead to assertion failures in
obscure cases)

-- A few mostly trivial performance tweaks. (Actually, inlining the default ctor
for Multiname is definitely worthwhile in Interpreter-only builds; the other
are mainly slight code-size savings.)

-- Not sure if it's possible for this to cause a problem in current Atom-only
builds, but in Box builds it was possibly for verifyEnter being used on a
function returning double to cause the result to get "lost" because the result
was in GPR rather than FPR result code. This splits verifyEnter into FPR/GPR
versions.

-- Global-scope rarely changes but we recalculate it every time. Caching it in a
local gives a modest but worthwhile win for code that is super-heavy on global
variable usage.

-- We can use the bitmask from AvmCore::istype to inline common cases for
Toplevel::coerce (ie, the cases where no coercion is necessary). This is
frequent enough to be a win for Interpreter-only builds.

-- Currently we require the caller to call Toplevel::coerce() prior to calling
setSlotAtom; setSlotAtom can and should do this internally, since many cases
(eg all numeric slots) will always succeed, and others are substantially
simplified. It also means the caller no longer needs to extract the proper slot
traits for coercion, which is doubly nice since setSlotAtom() already has that
handy internally and thus avoids a redundant call to getTraitsBindings.