nanojit x86 backend: rematerialize stack args instead of spilling (bug 513007 r=nnethercote+ sr=rreitmai+)
When we need to spill a LIns that is an incoming stack parameter, we
don't need to create new spill space for it - we can rematerialize it
directly from the stack, relative to EBP.

In tamarin methods, MethodEnv* env is a common argument that spills
a lot.  This will save one store in each method where that happens,
which sometimes will save one pointer-sized slot in the stack frame.