Remove unnecessary specialization of finddefNs and finddefNsset, clean up code (bug 515722 r=rreitmai+)
It might be a win to specialize finddefNs, however we aren't getting it
now because the two helpers just call MethodEnv::finddef.  Also I'm working
on adapting the wordcode interpreter's lookup cache to the jit, which will
eliminate most of the finddef() calls altogether.  this is the first patch
in that sequence.