Early bind slot load in OP_callprop (call to closure in slot) (bug 517888 r=rreitmai+)
When we encounter OP_callproperty on a known base type with a found
binding, but the binding is not a method, we currently don't optimize it.
Instead we should do a quick load (usually one instruction) and then call
the closure the normal way.

Some refactoring included, to make this maintainable:

* loading from slot is relegated to loadFromSlot() helper that doesn't update
FrameState, so it can be used as an intermediate step.  emitGetSlot() now calls
loadFromSlot in the obvious way.

* loadAtomRep factored to take Traits* as an argument, when the input value
isn't something sitting in a variable slot tracked in FrameState

* loadAtomRep cleaned up to switch builtin type instead of testing each type
one at a time.  removed some redundant constant folding (that is already
handled by nanojit::ExprFilter)

* op_call modified to call avmplus::op_call<MethodEnv*> which avoids loading
toplevel() in the common case when the function is not null or undefined.