Fix getprop_obj_slot<bool> on big endian machines (r=me)
Previous code incorrectly computed a Boolean slot pointer,
cast it to bool* and loaded a bool.  Since we store booleans
in int-sized slots, and sizeof(bool) == 1, we read the wrong
byte.  Solution is to use an enum Bool32 to make the slot int-sized.

we do that elsewhere in the thunk code too, and probably should
use something vm wide.