inline-cache for OP_setproperty (bug 520531 r=stejohns+)
Same as for getproperty and callproperty, with the additional twist that
a setproperty implies a property lookup *and* and implicit coercion.  When
the property type is a user class, the implicit coercion requires a Traits*
as an additional parameter, which in turn requires a bigger cache:

(function pointer, cached-vtable, slot-offset, coerce-type, Multiname*)

This patch includes a cleanup of the cache data types as well as
removing boilerplate for the fully late bound cases of setproperty & initproperty.