Extensively revamps the	-Dverbose flag.
verbose can now accept text fields following an equals sign ('='),
which provide further inputs to the verbosity mechanism.

So the following updates have been made:

  -Dverbose   - does what it used to (turns on full verbosity)
  -Dverbose_init - was 'broken' and is now removed.
  -Dverbose[=[parse,verify,interp,traits,builtins,jit,
              minaddr,memstats,sweep,occupancy]]

The 3rd form is the main addition in this change and allows
finer gain support for what is output.   The mechanism
consolidates all the 'verbose' bool's in the shell with a
single uint32_t which is encoded as a bitfield VB_Bits.

The -memstats and -memstats-verbose flags were removed
from the documentation, but are still functional at this point.

Their behaviour is now super-ceded by :
   -memstats         =>  -Dverbose=memstats,sweep
   -memstats-verbose =>  -Dverbose=memstats,sweep,occupancy

A fair amount of clean-up was also done, removing unused
fields and general cruft.

Finally, nanojit contains a very fine grained mechanism for
enabling verbosity.  The upper 16bits of the verbose_vb
field are now used to populate those flags.

This means that command-line -Dverbose= options can now
be propagated to nanojit.  Right now, only 'jit' and 'minaddr'
are being observed, but other fields can be easily supported
as needed.  (bug 506298)