Optimize OP_istypelate when RHS type is known (bug 516760 r=stejohns+ sr=lhansen+)
ASC currently only emits OP_istypelate for operator 'is', and usually the
value on the righthand side is statically known (a class or interface name).

This modifies the jit to inline the type constant, while preserving the evaluation
and null pointer check for the class value.  A future optimization or ASC change
may allow us to drop that evaluation and null-check, by using OP_istype or
optimizing OP_istypelate to be equivalent to OP_istype.

This change also allows calls to istype() to be optimized by CSE since they have
no side effects (unlike istypelate).