Bug 525347 - NJ merge: tamarin-specific nanojit-central import machinery [graydon] (r+me)
This patch carries scripts and files to automate imports from
nanojit-central to tamarin.

It works by noting revisions from nanojit-central, starting with the
rev number in the file nanojit-import-rev, and then pulling those
changes into the current repo.

The last step updates the rev number file and checks it in.

Note that the scripts will clobber the contents of the nanojit/
directory so it is important that changes move in a single
direction; i.e. from nanojit-central --> tamarin.

You also need to ensure that nothing nasty happened during this
import, and that you're the only person using the script when you
push your changes.

I would also recommend only running the script once per 'checkin',
and not try to revert or otherwise manipulate the repo after the
script has run.

If nanojit-import-rev (which identifies a node in nanojit-central)
gets corrupted (it really should never happen) you need to manually
set it to the correct rev, not doing so will cause subsequent
executions of this script to do very nasty things.

I found the following command quite handy in displaying
rev information: hg log --template "{rev} {node|short} {node}
{author}\n", since the long form of the node is used in the file.