Update to nanojit-central pull script
Add documentation to the script and support an option argument
which allows pulls from an alternate clone of a NJ repo.   Be
very careful when using this option as it might result in all kinds
of mayhem but it could be useful to pull a partial set of changes
from nanojit-central (rather than tip) if needed.