Emit LIR_live for every LIR_alloc passed to a call (bug 533280 r=rreitmai+)
In rare cases, it is possible for outgoing args to a function to be clobbered
when a register spills, because the code lifetime of the allocated space for
the args does not extend to the function call.  For example:

   args = LIR_alloc
   ...
   push /* some arg */
   lea eax, EBP+offset    /* address of args */
   /* spill here, clobbering args */
   call

Since the point-of-use of a LIR_alloc does not necessarily represent the
end of the lifetime of the underlying stack space, we use explicit LIR_live
instructions to demarcate the lifetime of the stack space itself.

HG Ener commit message.  Lines beginning with 'HG:' are removed.