Pass verifier-determined result types down into CodeWriter, and use them in the jit. (bug 546435 r=rreitmai+)
Pass verifier type information to CodeWriter so the jit can use it directly
instead of second guessing.

In a handful of cases, the JIT was not marking the result of an opcode with the
same type as the verifier.  This will cause confusion and crashes when
debugging, at best.

This patch is slightly bigger than necessary, for consistency's sake: several
additional opcodes were updated to pass type to CodeWriter, even though the jit
already was using the correct type.