Call resolveSignatures() before attempting to look up method binding. (bug 548805 r=stejohns+)
It currently does it unconditionally, after looking up the binding.  If the
type is not yet resolved, a link error in the Traits can cause
_buildTraitsBindings() to assert.  Calling resolveSignatures first allows the
link error to be reported normally via a VerifyError, and doesn't change
the fact that resolveSignatures() is still called at that point in the Verifier.