Bug 543898 - Traits resolution dependent on abc parsing order, not Domain hierarchy (r+edwsmith)
This third patch (3 of 4) folds together the query and add
operations on a trait into a single addUniqueXXX() method.

One subtle change in behaviour is that the call site from PoolObject
using the param_traits now does a query/add where previously simply
an add was performed.  This transform is correct in the sense that a
domain-wide search is performed prior to this call, so it is guaranteed
that the addUniqueTrait call will succeed.