Bug 546660 - Clean up mess around LINUX, AVMPLUS_LINUX, __linux__ (r+lhansen) [kpalac]
As Tanenbaum said, "Standards are great because there are so many to choose
from."  Anyway:

- AVMPLUS_LINUX is not known to any of the configuration scripts.

- The configure.py script defines "LINUX" in some situations, as well as
  AVMPLUS_UNIX.

- One presumes that __linux__ is a GCC-ism, which may or may not be defined
  by the RVCT or other compilers that may or may not be used on Linux.

There are some instances where we need something more specific than "Unix" or
"Posix", so there appears to be a legitimate need for a way of detecting Unix
(just like we have a need to detect Solaris, and we currently do).

I recommend that this is cleaned up along the lines of our Solaris detection.