Use ADD instead of OR for pointer tagging, to aid better code-gen. (bug 555255, r=rreitmai+)
Modifies the JIT frontend to use LIR_add instructions for pointer tagging,
instead of LIR_or instructions.  Adds can be folded with other adds, and also
can be rematerialized with LEA on x86/x64, unlike LIR_or.

There is a risk that tagging an already-tagged pointer now changes the value,
whereas 'or' would have left it unchanged.  However, doing such a thing would
only be hiding a latent bug, would only work if the existing tag were the same
one being or'd, and would be totally fragile in the face of changes to the Atom
tag values.  I reviewed the code carefully and I'm confident we aren't doing
any such illegal re-tagging.

The same concern applies if we change un-tagging from AND to SUB, which we
*will* want to do at some point, because it allows field-offset calculations to
be folded together with the untagging operation.