Bug 565184: Optimize exception handling and remove verification anomalies (r=edwsmith+)
This corrects Bug 565184 - Corrupt ABC can cause execution to fall of the end of JIT-ed method,
and recasts the exception dispatch mechanism in a form that results in more predictable behavior
when exception-handling code departs from the expected idiom, such as catch blocks entered via
fallthrough or branches.  The unboxing of the thrown value is done within the exception dispatch,
prior to entry to the handler, so that no code need be be generated to do so in the handler.
An attempt is made to push as much of the exceptional-case work into the helper functions, so as
to minimize the amount of generated "boilerplate" required in each method requiring a handler.
The handler boilerplate is not generated at all if there are no instructions that can throw
within a try block.