Bug 576265 - Eliminate OP_abs_jump, OR: Don't generate ABC for object initialization (r=stejohns+ lhansen+)
This patch completely removes OP_abs_jump.  Abc object initialization is now
customized to the execution mechanism.  Added an InitVisitor class to support
iterating over traits fields, used for initially validating the fields, and
later initializing them or compiling initializer code.

If a constructor method is interpreted, we first call a stub which iterates
over Traits and assigns fields to their default values (if nonzero).  This
is used for the wordcode and abc interpreter.  (We could further optimize
wordcode to generate special initializer instructions that are interpreted).

If a constructor method is JIT-compiled, the JIT generates code at the
start of the constructor function to initialize each field.

JIT'd code should see no change in performance; interpreted code might be
slightly slower.  On the other hand, startup might be faster since we don't
have to allocate and generate a custom chunk of bytecode for each class.

Most importantly, parsing ABC code is simplified;  Without OP_abs_jump,
all abc code is contiguous, and we are guaranteed to see exception handler
tables before parsing code, which simplifies the verifier, wordcode
translator, and the JIT.