Bug 538181 - Separate LIR generation from verification (r=wmaddox+)
Due to callbacks and access to public fields, CodegenLIR cannot currently be
driven as a CodeWriter by anything other than the Verifier.

This patch creates a much cleaner separation between verification and
LIR gen, whereby CodegenLIR may be driven by any class which implements a
minimal interface (CodegenDriver).

Description:

CodegenLIR accesses the verifier via the FrameState object passed through its
CodeWriter interface. CodegenLIR requires the verifier for the following:

1) To access constant frame information, e.g. Verifier::max_stack.

2) General util calls and codegen-driver callbacks, e.g.
Verifier::readBinding(), Verifier::getFrameState()

In the attached patch, a FrameState object does not carry a reference to the
verifier. Instead, each of the above requirements are satisfied as follows:

1) Constant frame information was moved into MethodSignature, if it doesn't
vary anywhere in the ABC code for a method, or is included in FrameState,
otherwise.

2) General util calls and codegen-driver callbacks are grouped into the
CodegenDriver interface. The verifier implements this interface.

3) canAssign() and readBinding() did not not depend on transient Verifier
state, so were moved to class Traits.