Bug 570521 - With minor changes, interpGPR() and interpFPR() become jit-only helpers (r=stejohns)
This patch fixes a benign bugin Function.apply/call; the bug was
allowing the interpreter to be invoked via a fast path that was
intended to only be used by jit/native methods.  With the bug fixed,
the only calls to interpFPR/GPR are from JIT code.  Any call to the
interpreter from C++ goes through coerceEnter and one of the invoke
methods.

So, the patch moves the interpGPR and interpFPR stubs into exec-jit.cpp
and appropriately #ifdefs the code that sets up pointers to these stubs.