bug 580603: selftest exercising the adversarial mode. (r=treilly,sr=lhansen)
This selftest is careful to save&restore the GCHeap fields it mucks
with, so it should compose more nicely with the other selftests in
place.

I also gathered data mapping levels of dispersiveness to the
particular heap-limits that they OOM at (its in the comments), which
may be of interest if we need to adjust the (somewhat arbitrary)
choices I made for those two parameters.