Bug 539094 - Code cleanup factored out of OSR branch (r=wsharp)
Introduced a VARSIZE constant (8) in CodegenLIR.h, use it instead of
the literal 8 in lots of places.

Factored out argSize() calculation.

Factored emitDebugEnter(), called by writePrologue(), and moved
initNotNull(), emitInitializers(), and emitDebugEnter() to after
we switch the LirWriter from prolog to body.  This shouldn't matter
now, but is needed when OSR lands.

Made subclasses of InitVisitor have virtual destructors.