Bug 617244 - nanojit: remove AvmCore (r=edwsmith) . . attachment 495753 - TR patch (against TR 5632:00830e7731c4) . Created attachment 495751 NJ patch (against TM 58520:8921e3faccd2)
Something I came across while working on bug 529430:  there's this AvmCore
class that is used throughout Nanojit.  It's a Tamarin-ism, AFAICT, and
isn't needed, and it's quite confusing.  All that *is* needed is the Config
struct that AvmCore (statically) contains.  So I decided to clean this up.

The NJ patch:

- Removes AvmCore from the nanojit/avmplus.h file that TM and lirasm use
  (but TR doesn't).  Removes all traces of AvmCore elsewhere.

- Removes some other dead stuff from avmplus.h.

- nInit() is now passed no arguments;  none of the back-ends need an argument.

- Removes 'has_cmov' from the Sparc back-end;  it is dead.

- Moves use_cmov() from AvmCore to Config.

- Updates lirasm for all the above changes, by adding a nanojit::Config to
  class Lirasm.