Add -Dverbose=lir mode to generate LIR-only listings (r=rreitmai+)
CodegenLIR::printNotNull() was extended to only print the notnull
status for valid scope/stack locations.  Locations above the stack
are not printed since they don't matter.

Interface ID's (IID) are always 0 in -Dverifyonly mode, to suppress
memory-layout noise.  (normally they are hashed pointers).

Only print deadvars listing if the LC_AfterDCE flag is set,
which will be the case with -Dverbose=jit, but not otherwise.