Bug 539094 - On-stack replacement, port from tr-spicy (r=edwsmith,rreitmai)
This is a cumulative patch containing the OSR work done in the tr-spicy branch.
The original work was reviewed piecemeal by edwsmith, and the rebased cumulative
patch by rreitmai.  OSR may be enabled with the --enable-osr configuration switch.
It is currently disabled by default.  While it has been used on a limited basis
in production, unresolved interpreter/jit semantic divergences make general
deployment inadvisable at this time. Nonetheless, we are required to carry this
functionality forward at this point in support of hosts that use it.