Bug 528375 - Add jit-min and jit-max compilation policy (r+kpalacz,edwsmith)
Created a framework for identifying methods via method name or unique method id - aka MethodRecognizer.

The first usage of this infrastructure is to allow one to override the default JIT compilation policy with your own rules.  The rules currently are read in from the command line via the -policy options (see help for details).