Bug 659380: check if bug is security tagged before push (author=cpeyer,r=fklockii).
The current tamarin preoutgoing_hook checks for MARK_SECURITY_CHANGE
in the diff being committed, but if does not actually check to see if
the bug is a marked as a security bug.

We should be able to check if it is a security bug by trying to get to
the bug as an anon user, which will not be allowed.

We unconditionally prompt "are you sure" for any push of a security
patch to any target, local or remote.  (We could attempt to narrow the
scope of the warning, but cpeyer convinced me we're better off being
annoying here.)