Bug 645878: Fix windows 32-bit builds (r=fklockii).
This is the error being fixed:

MMgcPortWin.cpp
..\..\AVMPI\MMgcPortWin.cpp(705) : error C2220: warning treated as error
- no 'object' file generated
..\..\AVMPI\MMgcPortWin.cpp(705) : warning C4826: Conversion from 'void
*' to 'unsigned __int64' is sign-extended. This may cause unexpected
runtime behavior.

There was a little bit of email discussion between Stan and Lars about
disabling the warning itself, but it seems easier to me to just put in
the extra cast.  I believe the cast correctly addresses the issue
described by the warning.