Bug 624164 - [NPOTB] complete PowerPC nanojit for Firefox (r=rreitmai sr=edwsmith) . . attachment 542715 - Part 2: PPC backend . User-Agent:       Mozilla/5.0 (Macintosh; PPC Mac OS X 10.4; rv:2.0b8pre) Gecko/20110107 Firefox/4.0b8pre TenFourFox/Debugging Build Identifier:
This adds code for Firefox nanojit support for PowerPC to the tree. This bug does NOT enable PPC nanojit in the build (thus NPOTB); that will be in a future patch. However, it is already enabled in TenFourFox and will be part of beta 9 of that browser. I discussed this work with Edwin Smith in E-mail already -- thanks for all your help, Ed!

This patch does the following:

- Adds nFragExit to NativePPC.cpp
- Adds support for the overflow math instructions to asm_arith() in NativePPC.cpp
- Completes nPatchBranch in NativePPC so that it can "demote" 14-bit and 24-bit branch displacements to CTR-based branches, and conversely "promote" CTR-branches to 14-bit displacements where possible, in NativePPC.cpp
- Adds additional opcodes to NativePPC.h required by the above
- Implements a basic instruction scheduler that hoists independent instructions up higher to facilitate better ILP. This is done by (ab)using the EMIT1 macro in NativePPC.h to do instruction swapping, and adding a new struct to Assembler that tracks instruction history. Special cases are added for certain common sequences that this scheduler does not catch, as well as macros for disabling the optimizer for sequences that must be emitted in strictly serialized order.

Performance is mixed. It actually takes a non-trivial hit on SunSpider to use the JIT, although it does significantly better on many parts of Dromaeo. The net effect is overall positive for most tasks, but not nearly enough. Therefore, I'm posting this first working draft to get more eyes on it and suggestions for further optimization.

Notes on stuff that came up during development and other related bugs Ed and Nick Nethercote pointed out in our E-mail conversation:

- There are lots of stack stores in hot code related to side exits (an empty loop takes an unbelievable amount of time because of this). This hurts badly on POWER which can take a big stall hit for memory access. The suggested solution for this is, as I understand it, controversial (bug 537842).

- Some of the load and store noise could be improved by bug 514102, and possibly further with bug 602793.

- Bug 545406 "TM: loop invariant code motion (LICM)" will probably benefit PPC greatly.

- For certain cases where we can statistically predict the PPC nanojit will take a bath with traced code, bug 597439 where we can let the compiler do our optimization (of the LIR interpreter itself) may also be useful. I have not made any tracer changes in this patch to that end, although I am exploring them.

Reproducible: Always

Steps to Reproduce:
n/a
Actual Results:
Warp 3.

Expected Results:
Ludicrous speed.