Tighten up aligned allocation in nanojit::Allocator.
Formerly, 8-byte alignment was maintaned by always rounding up the allocation
to keep the "next available" location on an 8-byte boundary.  Changes for float/float4
combined this approach with additional logic to allow allocations to be made with more
restrictive alignment, on a per-allocation basis.  This patch now adds all of the
necessary padding during the present allocation, and omits the now unnecessary rounding
logic that attempted to maintain alignment of the allocation pointer for the benefit of
the next allocation.