Bug 732918: add lzma project to xcode (r=fklockii).
I used a static library rather than a dynamic dylib.  I don't have
much reasoning behind this choice, it just seemed like the simplest to
fix the main problem here, and the code size here is hopefully not a
concern.  (Plus, IIUC, the Flash Runtime also uses static libraries
for targets like this, since in the end we are just producing a single
distributable.)

I made both 32- and 64-bit variants because that seemed like the
simplest way to deal with the linking dependency for the avm64
variant.  This may or may not represent a divergence from the
structure one should expect from the xplat builds.

I set 'GCC_WARN_ABOUT_MISSING_PROTOTYPES = NO;' because gcc is
complaining about the definition of MatchFinder_GetIndexByte in
LzFind.c without a prototype preceding it.  (This strikes me as a
somewhat weird thing to require; maybe good C style recommends one put
prototypes for all functions.  A counter-argument: I figure not
including a prototype is providing more information to the reader,
namely that this function is not referenced before its definition.)

Finally I added a newline to the end of LzmaEnc.c because the Xcode
gcc warns about such missing newlines, and I figure it is easier to
add them than to figure out how to change that in the configuration.