Bug 724139: Add new telemetry based sampler into VM (p=gcomnino, r pending=rulohani).
Cherry-picked from CL @1044455, though this is just the VM
changes (they continue to shelf-build on their own; the other
changes, to the outer context, will land with a pending FR
Main to FR Avm integration). The only difference from
conservative auto-resolve and auto-merge is that I replaced
some tabs with spaces in core/MethodInfo.h

fklockii has not regenerated the generated/builtin.* or
shell.* code himself.  That can be part of a followup
changelist if need be.

Command sequence:
% p4 integrate -1 -d -t main-code/third_party/avmplus/...@1044455,1044455 avm-code/third_party/avmplus/...
% p4 resolve -as avm-code/third_party/avmplus/...
% p4 resolve -n avm-code/third_party/avmplus/...
% p4 resolve -du avm-code/third_party/avmplus/core/AvmCore-inlines.h
# => Reviewed, then am
% p4 resolve -du avm-code/third_party/avmplus/core/MethodInfo.h
# => Reviewed, replaced tabs with spaces, then ae.
% p4 change

CL@1044829