No bug: Update the buildbot URL that is used in emails about failures in the deep phase (r=brbaker).
The only part of the build phase that is left in buildbot at the moment is the deep phase of the build system. At this stage we have disabled the external view of buildbot. This change modifies the URL that will be used in failure email to use the internal IP of the buildbot system.

Along with this change access to http://tamarin-builds.mozilla.org/tamarin-redux has been disabled.

{{mercurial 3f8be2a65bec6b042f80678d540edf4b57debec1}}

CL@1050083