Bug 741033: Support per-method control of execution policy including OSR opt-in (r=fklockii)
Recognizes a per-method metadata attribute 'ExecPolicy' while reading ABC.
An attribute of the form 'ExecPolicy(OSR=n)' allows per-method setting of the OSR threshold.
This also allows OSR opt-in for builds in which OSR is globally not enabled.
It is still possible to suppress OSR entirely, for diagnostic purposes, using the -osr=0 flag in
avmshell.

CL@1052155