No bug: Moving device testing information out of build script and into Jenkins (r=brbaker)
Currently the information about what devices are to be used for remote testing is contained within the source tree. This makes it difficult when we need to make a change for new devices becoming available or devices needing to go offline because the source change needs to be propagated to all of our active branches. With this change all that needs to be updated is the Jenkins configuration.

CL@1081843